+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 - App Metrics Batch fixed
+- Stats Batch fixed
 
 ## [1.0.4] - 2021-04-21